# docs/api/openapi.yaml
openapi: 3.0.3
info:
    title: Adblock Compiler API
    description: |
        **Compiler-as-a-Service** for adblock filter lists. Transform, optimize, and combine filter lists from multiple sources with real-time progress tracking.

        ## Features
        - 🎯 Multi-Source Compilation
        - ⚡ Performance (Gzip compression, caching, request deduplication)
        - 🔄 Circuit Breaker with retry logic
        - 📊 Visual Diff between compilations
        - 📡 Real-time progress via SSE and WebSocket
        - 🎪 Batch Processing
        - 🌍 Universal (Deno, Node.js, Cloudflare Workers, browsers)

        ## Links
        - [GitHub Repository](https://github.com/jaypatrick/adblock-compiler)
        - [Documentation](https://github.com/jaypatrick/adblock-compiler/tree/master/docs)
        - [Web UI](https://adblock-compiler.jayson-knight.workers.dev/)

    version: 2.0.0
    license:
        name: GPL-3.0
        url: https://github.com/jaypatrick/adblock-compiler/blob/master/LICENSE
    contact:
        name: Jayson Knight
        url: https://github.com/jaypatrick/adblock-compiler
servers:
    - url: https://adblock-compiler.jayson-knight.workers.dev
      description: Production server
    - url: http://localhost:8787
      description: Local development server
...
paths:
...
    /browser/resolve-url:
        post:
            tags: [Browser Rendering]
            summary: Resolve canonical URL
            description: |
                Resolves the canonical URL of a given address by following all redirects,
                including JavaScript-triggered ones, using a headless Chromium browser via
                Cloudflare Browser Rendering.

                Useful for determining the final destination of short links, CDN redirects,
                or mirror systems before caching filter list metadata.

                Returns **503** if the `BROWSER` binding is not configured.
            operationId: resolveCanonicalUrl
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UrlResolveRequest'
                        examples:
                            shortLink:
                                summary: Resolve a short link
                                value:
                                    url: https://short.link/abc
            responses:
                '200':
                    description: Canonical URL resolved successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UrlResolveResponse'
                '400':
                    $ref: '#/components/responses/BadRequestError'
                '502':
                    description: Navigation failed (unreachable URL, DNS error, etc.)
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: false
                                    error:
                                        type: string
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
...
components:
    schemas:
...
        UrlResolveRequest:
            type: object
            required:
                - url
            properties:
                url:
                    type: string
                    format: uri
                    description: The URL to resolve (follows all redirects including JS-triggered ones)
                    example: https://short.link/abc
        UrlResolveResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                canonical:
                    type: string
                    format: uri
                    description: The final canonical URL after all redirects
                    example: https://example.com/filters/list.txt
                hops:
                    type: integer
                    description: Number of URL changes observed during navigation
                    example: 2
...
    responses:
        BadRequestError:
            description: Bad request - request body failed schema validation
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            success:
                                type: boolean
                                example: false
                            error:
                                type: string
                                example: Invalid request body
...
        ServiceUnavailable:
            description: Service unavailable - required binding not configured
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            success:
                                type: boolean
                                example: false
                            error:
                                type: string
    securitySchemes:
        Turnstile:
            type: apiKey
            in: header
            name: turnstileToken
            description: Cloudflare Turnstile token (if enabled on the server)
        AdminKey:
            type: apiKey
            in: header
            name: X-Admin-Key
            description: Admin API key for protected admin endpoints
security: []